requirements.txt
openpyxl==3.1.2
pandas==2.2.0
psycopg2-binary==2.9.9
